Somewhat finished guard AI
Currently the guard seems to go into the search state, when first detecting the player. At least the the function to set the target to the last seen location is called, even when I'm right in front of the guard. Then, when the guard loses me, it doesn't even call that very same function again. Almost like it's doing the opposite of what it's supposed to do. Very wacky.

My theory is that is has something to do with the way I'm handling selectors, or routines. Maybe the abort system doesn't really work as intended. So maybe start looking there future me.